Tar høyde for minimal person og sjekker om det er infotrygdutbetalinger som skal slettes
Hvis delete-query opprettes når infotrygdutbetalingerRef er en tom liste vil det bli en syntaksfeil ved kjøring.

+++ spesialist-opprydding-dev/src/main/kotlin/no/nav/helse/opprydding/PersonRepository.kt
@@ -76,6 +76,7 @@ internal class PersonRepository(private val dataSource: DataSource) {
     }
 
     private fun TransactionalSession.slettInfotrygdutbetalinger(infotrygdutbetalingerRef: List<Int>) {
+        if (infotrygdutbetalingerRef.isEmpty()) return
         @Language("PostgreSQL")
         val query = "DELETE FROM infotrygdutbetalinger WHERE id IN (${infotrygdutbetalingerRef.joinToString { "?" }})"
         run(queryOf(query, *infotrygdutbetalingerRef.toTypedArray()).asExecute)
